Addresses #278
Supertypes and Subtypes, Mana Cost, and Sets are now stored as Lists of Strings. This fixes issues with tokens getting merged when saving and restoring options.
Handling those four fields is easier now without the need for concatenating or splitting Strings
SearchCriteria changed, which breaks previously stored settings, but I'm OK with users losing persisted search options on the upgrade to 3.4.3